Add presentation for half-dat intro
This presentation is an intentionally scaled-down version of the SYCL
Academy content, featuring slides from other lessons, compressed into
one slide deck for a faster introduction to SYCL. It is still worth
looking at all the content if time is not a pressure, but for shorter
sessions this might be more appropriate. It also covers the sample
which alters the convolution sample to have two 1D passes over the
input image.